Setup React
