+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^13.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "5.2",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.0"
   },
