+++ src/App.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+function App() {
+  return (
+  <></>
+  );
+}
+
+export default App;
